seperate display unify function to another class
