+++ Assets/Project/Scenes/GamePlayScene.unity
@@ -1497,6 +1497,51 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 1398619168}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &1427070992
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1427070994}
+  - component: {fileID: 1427070993}
+  m_Layer: 0
+  m_Name: DiplayUnifier
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1427070993
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1427070992}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5c4792476b1e64291b7ace017ce35779, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _backgroundPrefab: {fileID: 2028464485277968159, guid: ad43dc81975bb4cd4bada34c0925747f,
+    type: 3}
+--- !u!4 &1427070994
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1427070992}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 7
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1542554563
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1593,8 +1638,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   state: 0
-  _backgroundPrefab: {fileID: 2028464485277968159, guid: ad43dc81975bb4cd4bada34c0925747f,
-    type: 3}
 --- !u!4 &1638029897
 Transform:
   m_ObjectHideFlags: 0
